requirements.txt
# Sentra Tunnel Server - Central Management Server
# For remote access to Raspberry Pi door control systems

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# WebSocket Support
websockets>=12.0

# Database
aiosqlite>=0.19.0  # Async SQLite

# Templates
jinja2>=3.1.0

# Authentication
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4  # Password hashing

# HTTP Client (for testing)
httpx>=0.26.0

# Form handling
python-multipart>=0.0.6

# Configuration
pydantic-settings>=2.1.0

# Development
python-dotenv>=1.0.0
